requirements.txt
# Data Collection
requests==2.31.0
pandas==2.1.0
numpy==1.24.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Notebooks
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0